+++ Examples/Example-GetEmailMessage.ps1
@@ -92,6 +92,6 @@ $ClientSecret = 'your-client-secret'
 $TenantId = 'your-tenant-id'
 $cred = ConvertTo-GraphCredential -ClientId $ClientId -ClientSecret $ClientSecret -DirectoryId $TenantId
 $graph = Connect-EmailGraph -Credential $cred
-Get-EmailMessage -UserPrincipalName 'user@example.com' -Connection $graph -HasAttachment -Limit 5
+Get-GraphMessage -UserPrincipalName 'user@example.com' -Connection $graph -HasAttachment -Limit 5
 
 Disconnect-EmailGraph -Connection $graph
