+++ Examples/Example-GraphManageMessages.ps1
@@ -8,7 +8,7 @@ $TenantId = 'your-tenant-id'
 $cred = ConvertTo-GraphCredential -ClientId $ClientId -ClientSecret $ClientSecret -DirectoryId $TenantId
 $graph = Connect-EmailGraph -Credential $cred
 
-$messages = Get-EmailMessage -UserPrincipalName 'user@example.com' -Connection $graph -Filter "hasAttachments eq true" -Limit 5
+$messages = Get-GraphMessage -UserPrincipalName 'user@example.com' -Connection $graph -Filter "hasAttachments eq true" -Limit 5
 foreach ($m in $messages) {
     Get-MailMessageAttachment -UserPrincipalName 'user@example.com' -MessageId $m.Id -Connection $graph |
         Save-MailMessageAttachment -Path 'C:\Temp\Attachments'
