+++ Sources/Mailozaurr.PowerShell/CmdletGetEmailMessage.cs
@@ -1,10 +1,7 @@
 using MimeKit;
 using Mailozaurr;
 using System;
-using System.Collections.Generic;
-using System.Collections;
 using System.Management.Automation;
-using System.Management.Automation.Runspaces;
 using System.Threading.Tasks;
 
 namespace Mailozaurr.PowerShell;
@@ -43,30 +40,6 @@ public sealed class CmdletGetEmailMessage : AsyncPSCmdlet {
     [ValidateNotNull]
     public PopConnectionInfo? PopClient { get; set; }
 
-    [Parameter(Mandatory = true, ParameterSetName = "Graph")]
-    [Parameter(Mandatory = true, ParameterSetName = "MgGraphRequest")]
-    [ValidateNotNullOrEmpty]
-    public string? UserPrincipalName { get; set; }
-
-    [Parameter(Mandatory = true, ParameterSetName = "Graph")]
-    [ValidateNotNull]
-    public GraphConnectionInfo? Connection { get; set; }
-
-    [Parameter(ParameterSetName = "Graph")]
-    [Parameter(ParameterSetName = "MgGraphRequest")]
-    public string[]? Property { get; set; }
-
-    [Parameter(ParameterSetName = "Graph")]
-    [Parameter(ParameterSetName = "MgGraphRequest")]
-    public string? Filter { get; set; }
-
-    [Parameter(ParameterSetName = "Graph")]
-    [Parameter(ParameterSetName = "MgGraphRequest")]
-    public int? Limit { get; set; }
-
-
-    [Parameter(Mandatory = true, ParameterSetName = "MgGraphRequest")]
-    public SwitchParameter MgGraphRequest { get; set; }
 
 
     /// <summary>
@@ -131,19 +104,11 @@ public sealed class CmdletGetEmailMessage : AsyncPSCmdlet {
 
     /// <inheritdoc />
     protected override Task ProcessRecordAsync() {
-        switch (ParameterSetName) {
-            case "IMAP":
-                return ProcessImapAsync();
-            case "POP":
-                return ProcessPopAsync();
-            case "Graph":
-                return ProcessGraphAsync(Connection!.Credential);
-            case "MgGraphRequest":
-                ProcessMgGraph();
-                return Task.CompletedTask;
-            default:
-                return Task.CompletedTask;
-        }
+        return ParameterSetName switch {
+            "IMAP" => ProcessImapAsync(),
+            "POP"  => ProcessPopAsync(),
+            _      => Task.CompletedTask,
+        };
     }
 
     private Task ProcessImapAsync() {
@@ -192,90 +157,4 @@ public sealed class CmdletGetEmailMessage : AsyncPSCmdlet {
     }
 
 
-    private async Task ProcessGraphAsync(GraphCredential cred) {
-
-        var filters = new List<string>();
-        if (!All.IsPresent) {
-            if (!string.IsNullOrEmpty(Subject)) {
-                filters.Add($"contains(subject,'{Subject.Replace("'", "''")}')");
-            }
-            if (!string.IsNullOrEmpty(FromContains)) {
-                filters.Add($"contains(from/emailAddress/address,'{FromContains.Replace("'", "''")}')");
-            }
-            if (!string.IsNullOrEmpty(ToContains)) {
-                filters.Add($"toRecipients/any(r:contains(r/emailAddress/address,'{ToContains.Replace("'", "''")}'))");
-            }
-            if (Since.HasValue) {
-                filters.Add($"receivedDateTime ge {Since.Value.ToUniversalTime():yyyy-MM-ddTHH:mm:ssZ}");
-            }
-            if (Before.HasValue) {
-                filters.Add($"receivedDateTime le {Before.Value.ToUniversalTime():yyyy-MM-ddTHH:mm:ssZ}");
-            }
-        }
-        if (HasAttachment.IsPresent) {
-            filters.Add("hasAttachments eq true");
-        }
-        var filter = string.Join(" and ", filters);
-        if (!string.IsNullOrEmpty(Filter)) {
-            filter = string.IsNullOrEmpty(filter) ? Filter : $"{filter} and {Filter}";
-        }
-
-        var messages = await MicrosoftGraphUtils.GetMailMessagesAsync(
-            cred,
-            UserPrincipalName!,
-            Property,
-            filter,
-            Limit);
-
-        foreach (var msg in messages) {
-            WriteObject(PSObject.AsPSObject(msg));
-            if (Delete.IsPresent && msg.TryGetValue("id", out var idObj) && idObj is string id) {
-                await MicrosoftGraphUtils.DeleteMailMessageAsync(cred, UserPrincipalName!, id);
-            }
-        }
-    }
-
-
-    private void ProcessMgGraph() {
-        var filters = new List<string>();
-        if (!All.IsPresent) {
-            if (!string.IsNullOrEmpty(Subject)) {
-                filters.Add($"contains(subject,'{Subject.Replace("'", "''")}')");
-            }
-            if (!string.IsNullOrEmpty(FromContains)) {
-                filters.Add($"contains(from/emailAddress/address,'{FromContains.Replace("'", "''")}')");
-            }
-            if (!string.IsNullOrEmpty(ToContains)) {
-                filters.Add($"toRecipients/any(r:contains(r/emailAddress/address,'{ToContains.Replace("'", "''")}'))");
-            }
-            if (Since.HasValue) {
-                filters.Add($"receivedDateTime ge {Since.Value.ToUniversalTime():yyyy-MM-ddTHH:mm:ssZ}");
-            }
-            if (Before.HasValue) {
-                filters.Add($"receivedDateTime le {Before.Value.ToUniversalTime():yyyy-MM-ddTHH:mm:ssZ}");
-            }
-        }
-        if (HasAttachment.IsPresent) {
-            filters.Add("hasAttachments eq true");
-        }
-        var filter = string.Join(" and ", filters);
-        if (!string.IsNullOrEmpty(Filter)) {
-            filter = string.IsNullOrEmpty(filter) ? Filter : $"{filter} and {Filter}";
-        }
-
-        var query = new Dictionary<string, object>();
-        if (!string.IsNullOrEmpty(filter)) query["$filter"] = filter;
-        if (Property != null && Property.Length > 0) query["$select"] = string.Join(",", Property);
-        if (Limit.HasValue) query["$top"] = Limit.Value.ToString();
-        var uri = MicrosoftGraphUtils.JoinUriQuery("https://graph.microsoft.com/v1.0", $"/users/{UserPrincipalName}/messages", query);
-
-        var parameters = new Hashtable { { "Method", "GET" }, { "Uri", uri } };
-        var ps = System.Management.Automation.PowerShell.Create(RunspaceMode.CurrentRunspace);
-        ps.AddCommand("Invoke-MgGraphRequest");
-        ps.AddParameters(parameters);
-        var results = ps.Invoke();
-        foreach (var res in results) {
-            WriteObject(res);
-        }
-    }
 }
